UTT+UTTC: Changed file format to allow for more than 256 places with the same name CH: Used long long for size computations - allows more than 4GB path data and graph data CHC: blockcache uses long long for block loading. Fixes errors for files larger than 4GB

+++ plugins/contractionhierarchies/blockcache.h
@@ -72,7 +72,7 @@ public:
 		}
 
 		//load block (skip first block -> header )
-		_inputFile.seek( ( block + 1 ) * _blockSize );
+		_inputFile.seek( ( ( long long ) block + 1 ) * _blockSize );
 		_inputFile.read(_cache + free_block * _blockSize, _blockSize );
 		_blocks[free_block].id = block;
